+++ src/main/php/lang/reflection/AccessingFailed.class.php
@@ -1,7 +1,6 @@
 <?php namespace lang\reflection;
 
-use lang\XPException;
-
-class AccessingFailed extends XPException {
+class AccessingFailed extends TargetException {
+  const MESSAGE = 'Failed accessing';
 
 }